+++ packages/fiori-app-sub-generator/test/int/fiori-elements/expected-output/alp_v4_cap_typescript/app/alp_v4_cap_typescript/webapp/test/integration/pages/BooksList.ts
@@ -1,8 +1,0 @@
-export const actions = {};
-
-export const assertions = {};
-
-export default class ListReport {
-    actions = actions;
-    assertions = assertions;
-}