+++ packages/ui5-test-writer/src/index.ts
@@ -1,4 +1,4 @@
 export { generateOPAFiles } from './fiori-elements-opa-writer.js';
 export { generateFreestyleOPAFiles } from './fiori-freestyle-opa-writer.js';
-export { addVirtualTestConfig } from './utils/opaQUnitUtils.js';
+export { addVirtualTestConfig } from './utils/virtualOpaUtils.js';
 export type { OPAGenerationOptions } from './types.js';