+++ packages/generator-odata-downloader/CHANGELOG.md
@@ -1,5 +1,56 @@
 # @sap-ux/generator-odata-downloader
 
+## 0.3.19
+
+### Patch Changes
+
+#### Dependency Updates
+
+- Upgrade patch-level dependencies [[aed328d](https://github.com/SAP/open-ux-tools/commit/aed328da8a5c93e226c58e4d7dc14c7c82756259)]
+
+#### Workspace Updates
+
+- @sap-ux/odata-service-writer 1.0.8 → 1.0.9
+- @sap-ux/axios-extension 2.0.3 → 2.0.4
+- @sap-ux/btp-utils 2.0.2 → 2.0.3
+- @sap-ux/store 2.0.1 → 2.0.2
+- @sap-ux/odata-service-inquirer 3.0.15 → 3.0.16
+- @sap-ux/inquirer-common 1.0.13 → 1.0.14
+- @sap-ux/fiori-generator-shared 1.0.12 → 1.0.13
+- @sap-ux/telemetry 1.0.10 → 1.0.11
+- @sap-ux/ui5-config 1.0.3 → 1.0.3
+
+## 0.3.18
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/project-access 2.1.2 → 2.1.3
+- @sap-ux/axios-extension 2.0.3 → 2.0.3
+- @sap-ux/fiori-generator-shared 1.0.11 → 1.0.12
+- @sap-ux/inquirer-common 1.0.12 → 1.0.13
+- @sap-ux/mockserver-config-writer 1.0.6 → 1.0.7
+- @sap-ux/odata-service-inquirer 3.0.14 → 3.0.15
+- @sap-ux/odata-service-writer 1.0.7 → 1.0.8
+- @sap-ux/telemetry 1.0.9 → 1.0.10
+
+## 0.3.17
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/odata-service-inquirer 3.0.13 → 3.0.14
+
+## 0.3.16
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/odata-service-inquirer 3.0.12 → 3.0.13
+
 ## 0.3.15
 
 ### Patch Changes
@@ -12,7 +63,7 @@
 
 ## 0.3.14
 
-*Released: 2026-06-12T06:53:23Z*
+_Released: 2026-06-12T06:53:23Z_
 
 ### Patch Changes
 
@@ -24,7 +75,7 @@
 
 ## 0.3.13
 
-*Released: 2026-06-11T10:54:17Z*
+_Released: 2026-06-11T10:54:17Z_
 
 ### Patch Changes
 
@@ -36,7 +87,7 @@
 
 ## 0.3.12
 
-*Released: 2026-06-10T17:43:20Z*
+_Released: 2026-06-10T17:43:20Z_
 
 ### Patch Changes
 
@@ -44,7 +95,7 @@
 
 ## 0.3.11
 
-*Released: 2026-06-10T10:53:28Z*
+_Released: 2026-06-10T10:53:28Z_
 
 ### Patch Changes
 
@@ -52,7 +103,7 @@
 
 ## 0.3.10
 
-*Released: 2026-06-10T09:57:42Z*
+_Released: 2026-06-10T09:57:42Z_
 
 ### Patch Changes
 
@@ -69,7 +120,7 @@
 
 ## 0.3.9
 
-*Released: 2026-06-09T14:35:01Z*
+_Released: 2026-06-09T14:35:01Z_
 
 ### Patch Changes
 
@@ -84,7 +135,7 @@
 
 ## 0.3.8
 
-*Released: 2026-06-09T13:18:16Z*
+_Released: 2026-06-09T13:18:16Z_
 
 ### Patch Changes
 
@@ -95,7 +146,7 @@
 
 ## 0.3.7
 
-*Released: 2026-06-04T14:49:00Z*
+_Released: 2026-06-04T14:49:00Z_
 
 ### Patch Changes
 
@@ -103,7 +154,7 @@
 
 ## 0.3.6
 
-*Released: 2026-06-04T13:54:21Z*
+_Released: 2026-06-04T13:54:21Z_
 
 ### Patch Changes
 
@@ -119,7 +170,7 @@
 
 ## 0.3.5
 
-*Released: 2026-06-04T10:19:37Z*
+_Released: 2026-06-04T10:19:37Z_
 
 ### Patch Changes
 
@@ -135,7 +186,7 @@
 
 ## 0.3.4
 
-*Released: 2026-06-03T14:58:37Z*
+_Released: 2026-06-03T14:58:37Z_
 
 ### Patch Changes
 
@@ -155,7 +206,7 @@
 
 ## 0.3.3
 
-*Released: 2026-06-03T13:52:44Z*
+_Released: 2026-06-03T13:52:44Z_
 
 ### Patch Changes
 
@@ -172,7 +223,7 @@
 
 ## 0.3.2
 
-*Released: 2026-06-02T11:35:17Z*
+_Released: 2026-06-02T11:35:17Z_
 
 ### Patch Changes
 
@@ -183,7 +234,7 @@
 
 ## 0.3.1
 
-*Released: 2026-06-01T15:15:26Z*
+_Released: 2026-06-01T15:15:26Z_
 
 ### Patch Changes
 
@@ -199,7 +250,7 @@
 
 ## 0.3.0
 
-*Released: 2026-05-30T20:54:07Z*
+_Released: 2026-05-30T20:54:07Z_
 
 ### Minor Changes
 
@@ -250,7 +301,7 @@
 
 ## 0.2.0
 
-*Released: 2026-05-15T08:12:20Z*
+_Released: 2026-05-15T08:12:20Z_
 
 ### Minor Changes
 
@@ -258,7 +309,7 @@
 
 ## 0.1.2
 
-*Released: 2026-05-14T11:45:51Z*
+_Released: 2026-05-14T11:45:51Z_
 
 ### Patch Changes
 
@@ -266,7 +317,7 @@
 
 ## 0.1.1
 
-*Released: 2026-05-11T16:10:02Z*
+_Released: 2026-05-11T16:10:02Z_
 
 ### Patch Changes
 
@@ -274,7 +325,7 @@
 
 ## 0.1.0
 
-*Released: 2026-05-07T12:44:55Z*
+_Released: 2026-05-07T12:44:55Z_
 
 ### Minor Changes
 
@@ -282,7 +333,7 @@
 
 ## 0.0.10
 
-*Released: 2026-03-30T22:24:11Z*
+_Released: 2026-03-30T22:24:11Z_
 
 ### Patch Changes
 
@@ -290,7 +341,7 @@
 
 ## 0.0.9
 
-*Released: 2026-03-23T14:11:51Z*
+_Released: 2026-03-23T14:11:51Z_
 
 ### Patch Changes